Move analytics to util's folder
