+++ src/components/newdapp/index.js
@@ -19,7 +19,7 @@ import Proptypes from 'prop-types';
 import SelectedTemplate from './selectTemplate';
 import ProjectDetails from './projectDetails';
 import Templates from '../../templates';
-import { logEvent, Analytics } from "../../analytics";
+import { logEvent, Analytics } from "../../utils/analytics";
 import DappfileItem from '../projecteditor/control/item/dappfileItem';
 import JSZipUtils from 'jszip-utils';
